pyproject.toml
[project]
name = "knowledge-bot-telegram"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "ay", email = "artem.a.yats@gmail.com" }]
dependencies = [
    "tiktoken>=0.11.0",
    "qdrant-client>=1.15.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "fastembed>=0.7.3",
    "pymupdf4llm>=0.0.27",
    "sentence-transformers>=5.1.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.18.2",
    "ruff>=0.13.1",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_bot_telegram"]

[tool.mypy]
files = ["src/knowledge_bot_telegram"]
ignore_missing_imports = true
